fix(sglang): dispatch the SGLang refit without importing a policy worker
`_refit_sglang_dispatch` ran in the driver but imported
`megatron_policy_worker` (or `dtensor_policy_worker_v2`) to reach
`refit_sglang_{colocated,distributed}`. Those modules import
`megatron.bridge` and `nemo_automodel` at module scope, and the driver
environment has neither: `sglang` is declared mutually exclusive with
`mcore`, `fsdp`, `automodel` and `vllm` in `[tool.uv] conflicts`, so the
driver can never hold both the SGLang stack and a trainer backend.

Every SGLang refit therefore died on its first attempt with

    NRL_SGLANG_REFIT_FAILURE transfer=broadcast refit=1 \
        phase=prepare_or_transfer error_type=ModuleNotFoundError

on a 2-node GB200 run. Both transfer modes were affected, and
`SGLangColocatedWeightSynchronizer._refit_colocated` carried an
independent copy of the same import.

The four helpers are driver-side orchestration: they drive both actor
groups through the pause/transfer/quarantine protocol and call only
`policy` and `policy_generation` facades, so they were merely filed in
the wrong module. Move them verbatim to `nemo_rl/weight_sync/`, beside
the synchronizer that calls them, and repoint both dispatch sites. The
worker modules re-export the names so existing callers keep working.

Verified on 2 nodes x 4 GPUs with a driver environment built without
`mcore`: `NRL_SGLANG_REFIT_GROUP_READY world_size=5 engines=4`, three
successful refits, no failures, `train/loss` through step 3.

The new test parses the source rather than importing it, so it fails on
the offending commit even where every backend happens to be installed.
