+++ tests/unit/weight_sync/test_weight_synchronizer.py
@@ -347,9 +347,7 @@ class TestHTTPWeightSynchronizer:
 # SGLangColocatedWeightSynchronizer
 # ---------------------------------------------------------------------------
 
-_SGLANG_REFIT_DRIVER = (
-    "nemo_rl.models.policy.workers.dtensor_policy_worker_v2.refit_sglang_colocated"
-)
+_SGLANG_REFIT_DRIVER = "nemo_rl.weight_sync.dtensor_refit_sglang.refit_sglang_colocated"
 
 
 class TestSGLangColocatedWeightSynchronizer:
